Run only needed tests
