+++ .github/scripts/run-unit-tests.sh
@@ -18,7 +18,7 @@
 set -e
 set -x
 
-OPTS+="-pl !integration-tests,!:druid-it-tools,!:druid-it-image,!:druid-it-cases"
+OPTS+="-pl services"
 OPTS+=" -Dsurefire.failIfNoSpecifiedTests=false -P skip-static-checks -Dweb.console.skip=true"
 OPTS+=" -Djacoco.destFile=target/jacoco-${HASH}.exec"
